add api for phase 3

+++ core/src/main/java/com/canu/services/JobService.java
@@ -1,30 +1,29 @@
 package com.canu.services;
 
+import com.canu.dto.requests.AdminJobCancelRequest;
+import com.canu.dto.requests.RatingUserRequest;
 import com.canu.dto.requests.UpdateJobRequest;
 import com.canu.dto.requests.UpdateJobStatusRequest;
+import com.canu.dto.responses.AdminRatingDto;
 import com.canu.dto.responses.JobDto;
+import com.canu.dto.responses.RatingDto;
 import com.canu.exception.GlobalValidationException;
-import com.canu.model.CanUModel;
-import com.canu.model.JobModel;
-import com.canu.model.SkillSetModel;
-import com.canu.model.TagModel;
-import com.canu.repositories.CanURepository;
-import com.canu.repositories.JobRepository;
-import com.canu.repositories.SkillSetRepository;
-import com.canu.repositories.TagRepository;
+import com.canu.model.*;
+import com.canu.repositories.*;
 import com.canu.specifications.JobFilter;
+import com.canu.specifications.JobRatingFilter;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Slice;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 import java.util.Set;
@@ -44,6 +43,8 @@ public class JobService {
     final private SkillSetRepository skillSetRepo;
 
     final private SocketService socketSvc;
+
+    final JobReviewRepository jobReviewRepo;
 
     private static final Logger logger = LoggerFactory.getLogger(JobService.class);
 
@@ -107,9 +108,9 @@ public class JobService {
     }
 
     public Page<JobDto> showJobList(JobFilter filter, Pageable p) {
-        Slice<JobModel> jobEntities = jobRepo.findAll(filter, p);
+        Page<JobModel> jobEntities = jobRepo.findAll(filter, p);
         List<JobDto> jobList = jobEntities.stream().map(r -> new JobDto(r)).collect(Collectors.toList());
-        Page<JobDto> result = new PageImpl<>(jobList, p, jobRepo.count());
+        Page<JobDto> result = new PageImpl<>(jobList, p, jobEntities.getTotalElements());
         return result;
     }
 
@@ -123,13 +124,34 @@ public class JobService {
         }
     }
 
-    public void cancelJob(Long jobId) {
+    public void cancelJob(Long jobId, String reason) {
         UserDetails user = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
         CanUModel uUser = canURepo.findByEmail(user.getUsername());
         JobModel job = jobRepo.findById(jobId)
                               .orElseThrow(() -> new GlobalValidationException("Cannot find the job"));
         uUser.validatePrivilege(job.getCreationUser());
-        jobRepo.delete(job);
+        job.setReason(reason);
+        if(JobModel.JobStatus.PENDING.equals(job.getStatus())) {
+            jobRepo.delete(job);
+        } else {
+            job.setStatus(JobModel.JobStatus.REQUEST_CANCEL);
+            jobRepo.save(job);
+        }
+
+    }
+
+    public void cancelJobByAdmin(AdminJobCancelRequest request) {
+        JobModel job = jobRepo.findById(request.getJobId())
+                              .orElseThrow(() -> new GlobalValidationException("Cannot find the job"));
+        if(request.getIsApproval()){
+            job.setCancelStatus(null);
+            jobRepo.delete(job);
+        } else {
+            job.setCancelStatus(JobModel.CancelStatus.REJECTED);
+            jobRepo.save(job);
+        }
+        job.setAdminReason(request.getNote());
+
     }
 
     public void updateJob(UpdateJobRequest request) {
@@ -192,6 +214,82 @@ public class JobService {
         }
         job.setStatus(JobModel.JobStatus.COMPLETED);
         socketSvc.noticeCanUJobComplete(job);
+    }
+
+    public void ratingUser(RatingUserRequest request) {
+        UserDetails user = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        CanUModel canu = canURepo.findByEmail(user.getUsername());
+        JobModel job = jobRepo.findById(request.getJobId())
+                              .orElseThrow(() -> new GlobalValidationException(
+                                      "Cannot find job with id: " + request.getJobId()));
+        if (!JobModel.JobStatus.COMPLETED.equals(job.getStatus())) {
+            throw new GlobalValidationException("Cannot review for uncompleted job");
+        }
+        CanUModel reviewer;
+        CanUModel targetUser;
+        if (job.getCreationUser().getId().equals(canu.getId())) {
+            reviewer = canu;
+            targetUser = job.getRequestedUser();
+            job.setRating(request.getValue());
+        } else {
+            reviewer = canu;
+            targetUser = job.getCreationUser();
+        }
+
+        if(jobReviewRepo.existsDistinctByJobAndReviewerAndTarget(job, reviewer, targetUser)){
+            throw new GlobalValidationException("only rating one time");
+        }
+
+        JobReviewerModel model = JobReviewerModel.builder()
+                                                 .job(job)
+                                                 .reviewer(reviewer)
+                                                 .target(targetUser)
+                                                 .rating(request.getValue())
+                                                 .content(request.getContent())
+                                                 .build();
+
+        jobReviewRepo.save(model);
+        jobRepo.save(job);
+    }
+
+
+    public Object getRatingList(JobRatingFilter filter, Pageable p) {
+        UserDetails user = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        CanUModel canu = canURepo.findByEmail(user.getUsername());
+        filter.setUserId(canu.getId());
+        return buildRatingList(filter, p);
+    }
+
+    public Object getReviewList(JobRatingFilter filter, Pageable p) {
+        filter.setIsGetReview(true);
+        return buildRatingList(filter, p);
+    }
+
+    private Page<JobDto> buildRatingList(JobRatingFilter filter, Pageable p){
+        Page<JobReviewerModel> data = jobReviewRepo.findAll(filter, p);
+        List ratingList = new ArrayList();
+        for(JobReviewerModel review : data){
+            ratingList.add(new RatingDto(review));
+        }
+        Page<JobDto> result = new PageImpl<>(ratingList, p, data.getTotalElements());
+        return result;
+    }
+
+    public Object getAdminRatingList(JobFilter filter, Pageable p) {
+        filter.setStatus(JobModel.JobStatus.COMPLETED);
+        Page<JobModel> data = jobRepo.findAll(filter, p);
+        List ratingList = new ArrayList();
+        for(JobModel job : data){
+            if(job.getReviewers().size() > 0) {
+                if(job.getReviewers().stream().anyMatch(r ->job.getCreationUser().getId().equals(r.getReviewer().getId()))) {
+                    ratingList.add(new AdminRatingDto(job));
+                } else {
+                    ratingList.add(new AdminRatingDto(job.getReviewers().get(0)));
+                }
+            }
+        }
+        Page<JobDto> result = new PageImpl<>(ratingList, p, data.getTotalElements());
+        return result;
     }
 
 }
